main: include sys/param.c
The MIN macro is used as of f8f0271beb8c28a6ac95750f931a07337bbcfdbc,
but no header include was added for it. Looks like it was indirectly
included via something else, but changes to the OS X 10.12 SDK mean
the macro is no longer available. This adds an explicit include for the
header it comes from.

Fixes #25.