Avoid de-reference in _compareCFDisplayModes
Currently, the _compareCFDisplayModes method uses CGDisplayModeRef* as
the first and second parameters, in alignment with the
CFComparatorFunction type alias declaration expecting an
UnsafeRawPointer. But CGDisplayModeRef is already a pointer to
CGDisplayMode and doesn't need a reference again to work as a
comparator. Doing so also introduces extra referencing and dereferencing
that can be avoided.

Update the method signature to pass CGDisplayModeRef directly.