Fetch max supported mode for each monitor
Newer versions of MacOS use a scaling factor (like 2x, which is the
default for newer MacBooks), leading to the default mode being half the
native resolution of the display. In such cases, it would be desirable
to fetch the maximum resolution supported by the display (native
resolution).

Add a getMax option to fetch this information.